Update webpack-dev-server to fix security vulnerabilities
Summary:
Upgrading to latest version of `webpack-dev-server` fixes all of the security vulnerabilities listed by `yarn audit` (and Github's dependabot), which should also close a bunch of tasks we got about the problem.

The newest version of `webpack-dev-server` also had an API change, so needed to edit the webpack configurations to account for that change.

Reviewed By: felixtrz

Differential Revision: D46337039

fbshipit-source-id: c6124b5dcea133963fb8ef50cb8852dbad271e3c